feat: refactor exception package naming for consistency
