+++ src/main/java/br/app/pdz/api/model/exception/UserNotFoundException.java
@@ -1,4 +1,4 @@
-package br.app.pdz.api.model.exeption;
+package br.app.pdz.api.model.exception;
 
 import org.springframework.http.HttpStatus;
