+++ src/main/java/br/app/pdz/api/service/AuthService.java
@@ -6,9 +6,9 @@ import br.app.pdz.api.dto.SignUpRequest;
 import br.app.pdz.api.model.EnumRole;
 import br.app.pdz.api.model.Role;
 import br.app.pdz.api.model.User;
-import br.app.pdz.api.model.exeption.RoleNotFoundException;
-import br.app.pdz.api.model.exeption.UserAlreadyExistsException;
-import br.app.pdz.api.model.exeption.UserNotFoundException;
+import br.app.pdz.api.model.exception.RoleNotFoundException;
+import br.app.pdz.api.model.exception.UserAlreadyExistsException;
+import br.app.pdz.api.model.exception.UserNotFoundException;
 import br.app.pdz.api.repository.RoleRepository;
 import br.app.pdz.api.repository.UserRepository;
 import br.app.pdz.api.dto.UserDTO;
